+++ src/lib.rs
@@ -231,6 +231,31 @@
 //! | `true` | no | Expression is interpreted as `Value::Bool` |
 //! | `.34` | no | Expression is interpreted as `Value::Float` |
 //!
+//! ### [serde](https://serde.rs)
+//!
+//! This crate implements `serde::de::Deserialize` for its type `Node` that represents a parsed expression tree.
+//! The implementation expects a `string` as input.
+//! Example parsing with [ron format](docs.rs/ron):
+//!
+//! ```rust
+//! extern crate ron;
+//! use evalexpr::*;
+//!
+//! let mut configuration = HashMapConfiguration::new();
+//! configuration.insert_variable("five", 5);
+//!
+//! // In ron format, strings are surrounded by "
+//! let serialized_free = "\"five * five\"";
+//! match ron::de::from_str::<Node>(serialized_free) {
+//!     Ok(free) => assert_eq!(free.eval_with_configuration(&configuration), Ok(Value::from(25))),
+//!     Err(error) => {
+//!         // Handle error
+//!     },
+//! }
+//! ```
+//!
+//! With `serde`, expressions can be integrated into arbitrarily complex data.
+//!
 //! ## License
 //!
 //! This crate is primarily distributed under the terms of the MIT license.
@@ -239,19 +264,21 @@
 
 #![warn(missing_docs)]
 
+#[cfg(test)]
+extern crate ron;
 #[cfg(feature = "serde")]
 extern crate serde;
 
 mod configuration;
 pub mod error;
+#[cfg(feature = "serde")]
+mod feature_serde;
 mod function;
 mod interface;
 mod operator;
 mod token;
 mod tree;
 mod value;
-#[cfg(feature = "serde")]
-mod feature_serde;
 
 // Exports