+++ tests/integration.rs
@@ -407,3 +407,15 @@ fn test_shortcut_functions() {
         Ok(vec![Value::Int(3), Value::Int(3)])
     );
 }
+
+#[cfg(feature = "serde")]
+#[test]
+fn test_serde() {
+    let strings = ["3", "4+4", "21^(2*2)--3>5||!true"];
+
+    for string in &strings {
+        let manual_tree = build_operator_tree(string).unwrap();
+        let serde_tree: Node = ron::de::from_str(&format!("\"{}\"", string)).unwrap();
+        assert_eq!(manual_tree.eval(), serde_tree.eval());
+    }
+}